Refactor tests and change expiration-query.ts to expiration-date.ts
