+++ web-app/__tests__/test-query.test.ts
@@ -1,5 +1,5 @@
 import { PrismaClient } from '@prisma/client';
-import getExpiration from '../helpers/expiration-query';
+import getExpiration from '../helpers/expiration-date';
 import getQueryResult from '../prisma/test-helpers/query';
 
 const prisma = new PrismaClient();
@@ -12,16 +12,6 @@ afterAll(async () => {
 
   await prisma.$disconnect();
 });
-// TODO: REFACTOR
-it('should return expiration date for jane doe', async () => {
-  // jane has a 180-dayinactivity after issue date
-  await expect(getExpiration('janedoe.testnet', '2022-03-02')).resolves.toEqual('2022-08-29');
-});
-
-it('should return expiration date for john doe', async () => {
-  // johnn does not have any 180-dayinactivity after issue date
-  await expect(getExpiration('johndoe.testnet', '2022-03-02')).resolves.toEqual('2022-09-21');
-});
 
 it('should return query result for jane doe', async () => {
   const queryResult = await getQueryResult('janedoe.testnet', '2022-03-02');
@@ -29,14 +19,12 @@ it('should return query result for jane doe', async () => {
   expect(queryResult).toEqual(
     expect.arrayContaining([
       {
-        moment: expect.any(String),
-        diff_to_previous_activity: expect.any(Number),
+        moment: '2022-12-23T09:46:39+00:00',
+        diff_to_previous_activity: 296,
         has_long_period_of_inactivity: true,
       },
     ]),
   );
-
-  console.log('jane doe', { queryResult });
 });
 
 it('should return query result for john doe', async () => {
@@ -45,12 +33,25 @@ it('should return query result for john doe', async () => {
   expect(queryResult).toEqual(
     expect.arrayContaining([
       {
-        moment: expect.any(String),
+        moment: '2022-03-25T16:09:06+00:00',
         diff_to_previous_activity: null,
         has_long_period_of_inactivity: false,
       },
     ]),
   );
+});
 
-  console.log('john doe', { queryResult });
+it('should return expiration date as Last Activity Date - (diff_to_previous_activity - 180) for account with 180-day inactivity period', async () => {
+  // jane has a 180-day inactivity after issue date
+  // 296 - 180 = 116
+  // Last Activity: '2022-12-23'
+  // Expiration Date = '2022-12-23' - 116 days = '2022-08-29'
+  await expect(getExpiration('janedoe.testnet', '2022-03-02')).resolves.toEqual('2022-08-29');
+});
+
+it('should return expiration date as Last Activity Date + 180 for account with no 180-day inactivity period', async () => {
+  // john does not have any 180-day inactivity after issue date
+  // Last Activity: '2022-03-25'
+  // Expiration Date = '2022-03-25' + 180 days = '2022-09-21'
+  await expect(getExpiration('johndoe.testnet', '2022-03-02')).resolves.toEqual('2022-09-21');
 });
