+++ web-app/pages/api/cert/[imageFileName].ts
@@ -8,7 +8,7 @@ import { getNearAccountWithoutAccountIdOrKeyStoreForBackend } from '../../../hel
 import { height, populateCert, width } from '../../../helpers/certificate-designs';
 import { addCacheHeader } from '../../../helpers/caching';
 import { formatDate } from '../../../helpers/time';
-import getExpiration from '../../../helpers/expiration-query';
+import getExpiration from '../../../helpers/expiration-date';
 
 export const HTTP_ERROR_CODE_MISSING = 404;
 const svg = 'svg';
